Migrate non-generic helper functions
