+++ saas/rti/adapters/adapters.py
@@ -2,11 +2,12 @@ from __future__ import annotations
 
 import json
 import os
+import subprocess
 import time
 from abc import abstractmethod, ABC
 from enum import Enum
 from threading import Lock, Thread
-from typing import Optional
+from typing import Optional, IO, TextIO, AnyStr
 
 import saas.node
 from saas.cryptography.helpers import encrypt_file, decrypt_file
@@ -15,9 +16,9 @@ from saas.cryptography.rsakeypair import RSAKeyPair
 from saas.dor.blueprint import DORProxy
 from saas.dor.exceptions import IdentityNotFoundError
 from saas.dor.protocol import DataObjectRepositoryP2PProtocol
-from saas.exceptions import SaaSException
-from saas.helpers import write_json_to_file, read_json_from_file, generate_random_string, create_symbolic_link, \
-    validate_json
+from saas.exceptions import SaaSException, RunCommandError
+from saas.helpers import write_json_to_file, read_json_from_file, generate_random_string, validate_json
+from saas.keystore.assets.credentials import SSHCredentials
 from saas.logging import Logging
 from saas.nodedb.service import NetworkNode
 from saas.p2p.exceptions import PeerUnavailableError
@@ -36,6 +37,117 @@ class ProcessorState(Enum):
     BUSY = 'busy'
     STOPPING = 'stopping'
     STOPPED = 'stopped'
+
+
+def parse_stream(name: str, pipe: IO[AnyStr], file: TextIO = None, triggers: dict = None) -> None:
+    while True:
+        # read the line, strip the '\n' and break if nothing left
+        line = pipe.readline().rstrip()
+        logger.debug(f"parse_stream[{name}]\t{line}")
+        if not line:
+            break
+
+        # if we have a file
+        if file is not None:
+            file.write(line+'\n')
+            file.flush()
+
+        # parse the lines for this round
+        if triggers is not None:
+            for pattern, info in triggers.items():
+                if pattern in line:
+                    info['func'](line, info['context'])
+
+
+def monitor_command(command: str, triggers: dict, ssh_credentials: SSHCredentials = None, cwd: str = None,
+                    stdout_path: str = None, stderr_path: str = None) -> (list[str], list[str]):
+
+    # wrap the command depending on whether it is to be executed locally or remote (if ssh credentials provided)
+    if ssh_credentials:
+        a = ['sshpass', '-p', ssh_credentials.key] if ssh_credentials.key_is_password else []
+        b = ['-i', ssh_credentials.key] if not ssh_credentials.key_is_password else []
+        c = ['-oHostKeyAlgorithms=+ssh-rsa']
+
+        wrapped_command = [*a, 'ssh', *b, *c, f"{ssh_credentials.login}@{ssh_credentials.host}", command]
+
+    else:
+        wrapped_command = ['bash', '-c', command]
+
+    with open(stdout_path, 'x') as f_stdout:
+        with open(stderr_path, 'x') as f_stderr:
+            proc = subprocess.Popen(wrapped_command, cwd=cwd, stdout=subprocess.PIPE, stderr=f_stderr,
+                                    universal_newlines=True)
+            while proc.poll() is None:
+                parse_stream('stdout', proc.stdout, file=f_stdout, triggers=triggers)
+
+            logger.debug(f"process is done: stdout={stdout_path} stderr={stderr_path}")
+            proc.stdout.close()
+
+
+def run_command(command: str, ssh_credentials: SSHCredentials = None,
+                suppress_exception: bool = False) -> subprocess.CompletedProcess:
+
+    # wrap the command depending on whether it is to be executed locally or remote (if ssh credentials provided)
+    if ssh_credentials:
+        a = ['sshpass', '-p', ssh_credentials.key] if ssh_credentials.key_is_password else []
+        b = ['-i', ssh_credentials.key] if not ssh_credentials.key_is_password else []
+        c = ['-oHostKeyAlgorithms=+ssh-rsa']
+
+        wrapped_command = [*a, 'ssh', *b, *c, f"{ssh_credentials.login}@{ssh_credentials.host}", command]
+
+    else:
+        wrapped_command = ['bash', '-c', command]
+
+    # execute command
+    result = subprocess.run(wrapped_command, capture_output=True)
+    if not suppress_exception and result.returncode != 0:
+        raise RunCommandError({
+            'wrapped_command': wrapped_command,
+            'ssh_credentials': ssh_credentials.record if ssh_credentials else None,
+            'result': result
+        })
+    return result
+
+
+def scp_local_to_remote(local_path: str, remote_path: str, ssh_credentials: SSHCredentials) -> None:
+    # generate the wrapped command
+    a = ['sshpass', '-p', ssh_credentials.key] if ssh_credentials.key_is_password else []
+    b = ['-i', ssh_credentials.key] if not ssh_credentials.key_is_password else []
+    c = ['-oHostKeyAlgorithms=+ssh-rsa']
+    wrapped_command = [*a, 'scp', *b, *c, local_path, f"{ssh_credentials.login}@{ssh_credentials.host}:{remote_path}"]
+
+    # execute command
+    result = subprocess.run(wrapped_command, capture_output=True)
+    if result.returncode != 0:
+        raise RunCommandError({
+            'wrapped_command': wrapped_command,
+            'ssh_credentials': ssh_credentials.record,
+            'result': result
+        })
+
+
+def scp_remote_to_local(remote_path: str, local_path: str, ssh_credentials: SSHCredentials) -> None:
+    # generate the wrapped command
+    a = ['sshpass', '-p', ssh_credentials.key] if ssh_credentials.key_is_password else []
+    b = ['-i', ssh_credentials.key] if not ssh_credentials.key_is_password else []
+    c = ['-oHostKeyAlgorithms=+ssh-rsa']
+    wrapped_command = [*a, 'scp', *b, *c, f"{ssh_credentials.login}@{ssh_credentials.host}:{remote_path}", local_path]
+
+    # execute command
+    result = subprocess.run(wrapped_command, capture_output=True)
+    if result.returncode != 0:
+        raise RunCommandError({
+            'wrapped_command': wrapped_command,
+            'ssh_credentials': ssh_credentials.record,
+            'result': result
+        })
+
+
+def create_symbolic_link(link_path: str, target_path: str, working_directory: str = None) -> None:
+    if working_directory:
+        link_path = os.path.join(working_directory, link_path)
+        target_path = os.path.join(working_directory, target_path)
+    run_command(f"ln -sf {target_path} {link_path}")
 
 
 class RTIProcessorAdapter(Thread, ABC):
